+++ mav_trajectory_generation/include/mav_trajectory_generation/io.h
